Provide Voikko as a bundled extension
...to fix <https://github.com/flathub/org.libreoffice.LibreOffice/issues/122>
"Add Voikko for Finnish spell check".

This is based on (and a partial copy of)
<https://github.com/jpakkane/org.libreoffice.LibreOffice/commit/b98d202f4deb6f7f13a27bf0b770dfe73b1de2f5>
"Add Finnish spell checker and hyphenation. Closes flathub#122." (cf.
<https://github.com/flathub/org.libreoffice.LibreOffice/pull/155> "Add Finnish
spell checker and hyphenation. Closes #122."), but provides its functionality as
a bundled extension rather than as a core part of the
org.libreoffice.LibreOffice app.  Also, in line with Fedora's
<https://src.fedoraproject.org/rpms/libvoikko/c/374ff8b09fd2ff289e7777ab22dcd21bce4ee1f5>
"Update to 4.3" (as used by Fedora's libreoffice-voikko equivalent of this
bundled flatpak extension), --disable-hfst to avoid a bunch of dependencies.

This relies on availability of
<https://github.com/flathub/org.libreoffice.LibreOffice/pull/158> "Add an
org.libreoffice.LibreOffice.BundledExtension extension point" in the
org.libreoffice.LibreOffice app.